Extract LovedOneThankYouRedesignedComponent
